% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_icd_10}
\alias{data_icd_10}
\title{Deaths due to natural events in Brazil from 1996 to 2019.}
\format{
A data frame with XXXX rows and 12 variables:
\describe{
\item{age_unit}{Units of age}
\item{age_value}{Age}
\item{birth_date}{Date of birth}
\item{cause}{Cause of death}
\item{residence_city}{Code of the city of residence}
\item{color_race}{Skin color or ethnicity}
\item{death_city}{Code of the city of death}
\item{death_date}{Date of the death}
\item{education}{Education level}
\item{job}{Code of profession}
\item{literacy}{Number of years of education}
\item{locus}{Place of death}
\item{marital}{Marital status}
\item{sex}{Sex}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
data_icd_10
}
\description{
A dataset containing information of causes of death caused by exposure to
forces of nature from 1996 to 2019. This data is based on death certificates
that capture more than 50 variables. Original source: World Health
Organization - International Statistic Classification of Diseases and Related
Health Problems 10th revision ICD10.
}
\keyword{datasets}
