% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_icd_9}
\alias{data_icd_9}
\title{Deaths due to natural events in Brazil from XXXX to 1995.}
\format{
A data frame with XXXX rows and 10 variables:
\describe{
\item{age_unit}{Units of age}
\item{age_value}{Age}
\item{birth_date}{Data of birth}
\item{cause}{Cause of death}
\item{residence_city}{Code of the city of residence}
\item{death_city}{Code of the city of death}
\item{death_date}{Date of the death. If the day of the month is unavailable, it is assumed the first day of the month}
\item{education}{Education level}
\item{job}{Code of profession}
\item{locus}{Place of death}
\item{marital}{Marital status}
\item{sex}{Sex}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
data_icd_9
}
\description{
TODO: Add description.
}
\keyword{datasets}
