% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazildatamortality.R
\name{get_data}
\alias{get_data}
\title{Merge the data comming from ICD 9 and ICD 10.}
\usage{
get_data()
}
\value{
A tibble with mortality data of Brazil.
}
\description{
Merge the data comming from ICD 9 and ICD 10.
}
