% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_hex}
\alias{state_hex}
\title{Hexagonal representation of the states of Brazil.}
\format{
A data frame with 27 rows and 6 variables:
\describe{
\item{code_state}{Code of the state}
\item{abbrev_state}{Abbreviated name of the state}
\item{name_state}{Name of the state}
\item{code_region}{Code of the state's region}
\item{name_region}{Name of the state's region}
\item{geometry}{Geometry}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
state_hex
}
\description{
Brazilian states (see state_sf) represented as hexagons using simple
features.
}
\keyword{datasets}
