% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_sf}
\alias{state_sf}
\title{States of Brazil as sf simple features.}
\format{
A data frame with 27 rows and 6 variables:
\describe{
\item{code_state}{Code of the state}
\item{abbrev_state}{Abbreviated name of the state}
\item{name_state}{Name of the state}
\item{code_region}{Code of the state's region}
\item{name_region}{Name of the state's region}
\item{geom}{Geometry}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
state_sf
}
\description{
Spatial limits of the Brazilian states in 2010 (simplified). These data was
obtained from the package geobr. For more detail see \url{https://github.com/ipeaGIT/geobr}.
}
\keyword{datasets}
