% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{town_sf}
\alias{town_sf}
\title{Towns of Brazil as sf simple features.}
\format{
A data frame with 5567 rows and 9 variables:
\describe{
\item{code_muni}{Code of the municipality}
\item{name_muni}{Name of the municipality}
\item{code_state}{Code of the sate}
\item{abbrev_state}{Abbreviated name of the state}
\item{code_meso}{Code of the meso-region}
\item{name_meso}{Name of the meso-region}
\item{code_micro}{Code of the micro-region}
\item{name_micro}{Name of the micro-region}
\item{name_state}{Name of the state}
\item{code_region}{Code of the region}
\item{name_region}{Name of the region}
\item{geom}{Geometry}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
town_sf
}
\description{
Spatial limits of the Brazilian towns in 2010 (simplified). These data was
obtained from the package geobr. For more detail see \url{https://github.com/ipeaGIT/geobr}.
}
\keyword{datasets}
